Remove hardcoded title wording
Search results titles had hardcoded strings.
They were not translated as at this stage of dev the i18n strategy
and serarch engine logic are both still not finished, and migration
might be tedious.

As the buttons hace an icon associated with the encountered element
it is understood which type of element was found.