:warning: Routing refactor (PHASE I)
This change impacts on almost all routes.
All businesses are now referenced by their slug right after the domain
root. This enables easy identification of business and lets the user know
on which business is he working at.

Tests are passing but further acceptance testing is needed at this point.