Disable "profile-reuse" mode
Profile reuse from not-owned businesses is disabled for privacy reasons,
as a shared profile would also share the updated data and would cause
race-conditions.

A profile may be reused only when it is already registered for the same
business, this way avoiding duplication.

A profile may be copied from another business to avoid user input of
already existing data.