Reorganize assets in purpose groups
Not all the assets are used together among all the app. In most of the
cases they are required as groups. Say a form view is loaded, then
those plugins providing form and input facilities need to be loaded,
and some others groups which are not needed may be avoided.