Remove Link User To Contact function
To be re-implemented.

This function represents a security issue as malicious user
may get registered with fake email in order to get linked to existing
contacts with that email address, thus exposing a contact profile
to unauthorized.

Linking should be performed right after registered user has confirmed
the ownership of that email address.

Inverse linking is still valid: A Contact gets linked to existing
user. The risk of "giving it out" to unauthorized is way less. However
this linking should also be performed under an email confirmation flag
check.