Add feature to allow custom timeslot_step
This allows a business to set a timeslot availability granularity, 
letting the client to choose a more comfortable hour for booking.
Enabling this suits better for certain categories of business, for 
example consulting. The default behavior uses the service 
duration for steps, probably fitting better for courses with a fixed
timetable.