Move captcha on top
In some devices like Smart TV, the form will be automatically
submitted after filling out the last input field, thus skipping
the captcha control and getting unable to succeed the registration.
Solving the captcha upfrond may become annoying for some users, so
it may be taken back to bottom area if a better workaround is found.