Create ISSUE_TEMPLATE.md
Hi, I am very interested in this project. I would like to see it continuing and I want to participate !
One of the best way is to allow people to send their feedback, their ideas. Those feedbacks have to be public in order to get anybody working on it. So we need an Issue section for this repo.
Please, allow us to keep it continuing, adding new features that could interest you too.