Fix business pref loading of opening and closing hours
Problem: The vacancies default time range is being taken from Business preferences, however when the time is '12:00 PM', it is loaded in the time picker as '12:00 AM', causing unwanted effects later on.

Fix: Load the time preference as a Carbon instance and re-formatting the value accordingly before passing to the view.

Fixes https://github.com/timegridio/timegrid/issues/196